feat(config): WithConfig mixin — block-agnostic config hosting
Extract the <uc-config> element-ConfigController adapter into a reusable
WithConfig(Base) mixin (src/lit): any ChildBlock can now host config the way
events/api already attach to any block. <uc-config> becomes a thin
'class Config extends WithConfig(ChildBlock)'. The config type surface
(instance props + attributesMeta) is auto-derived from ConfigType via ConfigHost,
replacing the hand-declared interface. One config host per ctx is enforced with
a deferred-confirm console warning (Policy 2) via the ConfigController writer
registry; a normal host swap stays silent. Behavior is byte-identical: the 27
Config specs pass unchanged and WithConfig.ts carries the moved coverage
(99.5% stmts / 100% lines). Module-level key data moved to config-keys.ts;
exported.ts repointed there.
