+++ src/lit/WithConfig.ts
@@ -1,0 +1,668 @@
+import { ConfigController } from '../abstract/controllers/ConfigController';
+import type { ControllerContainer } from '../abstract/di/ControllerContainer';
+import { inject } from '../abstract/di/inject';
+import { PluginManagerBridge } from '../abstract/di/PluginManagerBridge';
+import { lazy } from '../abstract/logger';
+import type { PluginController } from '../abstract/managers/plugin';
+import { runAssertions } from '../blocks/Config/assertions';
+import { type ComputedPropertyControllers, computeProperty } from '../blocks/Config/computed-properties';
+import { allConfigKeys, builtinAttrKeyMapping, isComplexKey, plainConfigKeys } from '../blocks/Config/config-keys';
+import { initialConfig } from '../blocks/Config/initialConfig';
+import { normalizeConfigValue } from '../blocks/Config/normalizeConfigValue';
+import type { ConfigType } from '../types';
+import { toKebabCase } from '../utils/toKebabCase';
+import type { ChildBlock } from './ChildBlock';
+import type { Constructor } from './Constructor';
+import type { ConfigHost } from './config-host-types';
+import { subscription, type Unsubscribe } from './subscription';
+
+/**
+ * Render a config value for the change log with consistent quoting — strings
+ * (incl. the empty string, which would otherwise be invisible) show quoted as
+ * `""`, functions as `ƒ`, and everything else via `JSON.stringify`.
+ */
+const formatConfigLogValue = (value: unknown): string => {
+  if (value === undefined) return 'undefined';
+  if (typeof value === 'function') return 'ƒ';
+  try {
+    return JSON.stringify(value) ?? String(value);
+  } catch {
+    // `String(value)` can ALSO throw (a throwing `Symbol.toPrimitive`/`toString`),
+    // so guard it too — logging a config change must never surface an exception.
+    try {
+      return String(value);
+    } catch {
+      return '[unserializable]';
+    }
+  }
+};
+
+const getLocalPropName = (key: string) => `__${key}`;
+
+/**
+ * Makes its `Base` block a **config host**: the element↔`ConfigController`
+ * adapter that used to live only on `<uc-config>`, now reusable on any
+ * `ChildBlock`. The host reads/writes attributes and DOM properties for every
+ * built-in config option (and dynamic plugin-registered custom configs) into the
+ * ctx's `ConfigController` — the SAME instance the rest of the app reads — and
+ * reflects external controller changes back onto its attributes/properties.
+ *
+ * It is class-level (not a bare reactive controller) so it can contribute the
+ * auto-inferred config type surface (`ConfigHost`, derived from `ConfigType`),
+ * `static observedAttributes`, and the property accessors onto the host's own
+ * type and prototype. `this` stays the element throughout — behavior is
+ * identical to the previous `<uc-config>` implementation.
+ *
+ * The plugin-manager (custom-config registry) reads go through the editor-safe
+ * `PluginManagerBridge` DI token rather than a direct `use(PluginController)`,
+ * for two reasons: (1) a config host can run in config-only ctxs where no
+ * uploader scope ever binds `PluginController`, so the reads must stay
+ * null-/absence-tolerant (`getOrNull`/`whenController` on an unbound token stay
+ * inert, where a synchronous `use(PluginController)` would throw); and (2)
+ * `<uc-config>` is in the editor-alone bundle, so a value import of
+ * `PluginController` here would drag it (and `PluginRegistry`) into that bundle
+ * and blow its size-limit — the `declare`-only `PluginManagerBridge` keeps it out
+ * while resolving to the SAME container instance.
+ */
+export function WithConfig<T extends abstract new (...args: any[]) => ChildBlock>(
+  Base: T,
+): T & Constructor<ConfigHost> {
+  abstract class WithConfigClass extends Base {
+    // Resolves the ctx's `ConfigController` via the container this block adopted
+    // (tagged as `this[CONTAINER]`), yielding the very same controller instance
+    // the ctx owns — writes land where every reader looks.
+    @inject(ConfigController) private readonly _config!: ConfigController;
+
+    private _computationControllers: ComputedPropertyControllers = new Map();
+    private _mutationObserver?: MutationObserver;
+    // The ctx's `ConfigController` captured at adoption. Used to deregister the
+    // writer in `controllerReleased` and inside the deferred multi-writer check,
+    // both of which run AFTER `_releaseController` clears `this[CONTAINER]` — so
+    // the `@inject _config` getter would throw there. The captured reference is
+    // the exact instance the writer registered on; `unregisterWriter` on an
+    // already-destroyed controller is a harmless no-op.
+    private _writerConfig?: ConfigController;
+
+    /**
+     * Mapping of attribute names to custom config keys for plugin-registered configs.
+     * Updated dynamically when plugins are registered or changed.
+     * Similar to builtinAttrKeyMapping but for custom configs.
+     */
+    private _customAttrKeyMapping: Record<string, string> = {};
+
+    /** Set of all custom config names registered by plugins */
+    private _customConfigKeys: Set<string> = new Set();
+
+    /**
+     * Map of custom config subscriptions (config name -> unsubscribe function)
+     * Used to track and clean up subscriptions when plugins change
+     */
+    private _customConfigSubscriptions: Map<string, () => void> = new Map();
+
+    /**
+     * Check if a key is a custom config (registered by plugins)
+     */
+    private _isCustomConfig(key: string): boolean {
+      return this._customConfigKeys.has(key);
+    }
+
+    /**
+     * Get the custom config definition for a key
+     */
+    private _getCustomConfigDefinition(key: string) {
+      // Read null-tolerantly via the `PluginManagerBridge` token: `getOrNull`
+      // yields `null` when no uploader scope has bound the bridge (config-only /
+      // plugin-less ctx) or when no container is adopted yet, instead of throwing.
+      // This helper is reachable with a stale `_customConfigKeys`/
+      // `_customAttrKeyMapping` — e.g. the MutationObserver forwards a custom
+      // attribute during a live `ctx-name` switch, before `attributeChangedCallback`
+      // reaches its own guard, or after a re-adoption into a ctx that has no plugin
+      // manager. (`getOrNull` — not `useOrNull` — because the bridge is a
+      // conditionally-bound token: a plain `get` on the unbound token would build a
+      // bogus zero-arg instance.)
+      const pluginManager = this.containerOrNull?.getOrNull(PluginManagerBridge)?.getPluginManager() ?? null;
+      if (!pluginManager) return undefined;
+      return pluginManager.configRegistry.get(key);
+    }
+
+    /**
+     * Get attribute names for a config key (kebab-case and lowercase)
+     */
+    private _getAttributeNames(key: string): string[] {
+      return [...new Set([toKebabCase(key), key.toLowerCase()])];
+    }
+
+    private _flushValueToAttribute(key: string, value: unknown) {
+      // Check if it's a complex built-in key
+      if (isComplexKey(key as keyof ConfigType)) {
+        return; // Complex keys can't be represented as attributes
+      }
+
+      // Check if it's a custom config with attribute: false
+      if (this._isCustomConfig(key)) {
+        const config = this._getCustomConfigDefinition(key);
+        // Skip if attribute is explicitly false (default is true, so flush unless false)
+        if (config?.attribute === false) {
+          return;
+        }
+      }
+
+      // Flush the value to the DOM attributes (works for both built-in and custom configs)
+      const attrs = this._getAttributeNames(key);
+      for (const attr of attrs) {
+        if (typeof value === 'undefined' || value === null) {
+          this.removeAttribute(attr);
+        } else if (this.getAttribute(attr) !== value.toString()) {
+          this.setAttribute(attr, value.toString());
+        }
+      }
+    }
+
+    private _flushValueToState(key: string, value: unknown) {
+      const isCustom = this._isCustomConfig(key);
+      const configKey = key as keyof ConfigType;
+      const currentValue = isCustom ? this._config.getCustom(key) : this._config.get(configKey);
+
+      if (currentValue !== value) {
+        if (typeof value === 'undefined' || value === null) {
+          // For built-in configs, use initial value; for custom configs, keep undefined
+          const defaultValue = initialConfig[configKey];
+          const nextValue = defaultValue !== undefined ? defaultValue : value;
+          if (isCustom) {
+            this._config.setCustom(key, nextValue);
+          } else {
+            this._config.set(configKey, nextValue as ConfigType[typeof configKey]);
+          }
+        } else {
+          if (isCustom) {
+            this._config.setCustom(key, value);
+          } else {
+            this._config.set(configKey, value as ConfigType[typeof configKey]);
+          }
+        }
+      }
+    }
+
+    private _setValue(key: string, value: unknown) {
+      // Normalize value (works for both built-in and custom configs)
+      let normalizedValue: unknown;
+
+      if (this._isCustomConfig(key)) {
+        // For custom configs, try to get normalize function from plugin definition
+        const config = this._getCustomConfigDefinition(key);
+        try {
+          normalizedValue = config?.normalize?.(value) ?? value;
+        } catch (error) {
+          this._log.warn(`normalize() for "${key}" threw an error, keeping previous value`, error);
+          return;
+        }
+      } else {
+        // For built-in configs, use the standard normalization
+        normalizedValue = normalizeConfigValue(key as keyof ConfigType, value);
+      }
+
+      // Perform the actual value setting
+      const anyThis = this as any;
+      const localPropName = getLocalPropName(key);
+
+      if (anyThis[localPropName] === normalizedValue) return;
+
+      this._assertSameValueDifferentReference(key, anyThis[localPropName], normalizedValue);
+
+      anyThis[localPropName] = normalizedValue;
+
+      // Flush the value to the state and attribute
+      this._flushValueToAttribute(key, normalizedValue);
+      this._flushValueToState(key, normalizedValue);
+      // NB: config-change logging is NOT done here — it's a separate observer on
+      // `ConfigController` set up in `controllerReady` (`_setupChangeLog`), so it
+      // catches every change (incl. non-block writers) decoupled from the setter.
+
+      // Only run assertions for built-in configs
+      if (!this._isCustomConfig(key)) {
+        runAssertions(this._config.values);
+      }
+    }
+
+    private _getValue(key: string) {
+      const anyThis = this as any;
+      const localPropName = getLocalPropName(key);
+      return (
+        anyThis[localPropName] ??
+        (this._isCustomConfig(key) ? this._config.getCustom(key) : this._config.get(key as keyof ConfigType))
+      );
+    }
+
+    private _assertSameValueDifferentReference(key: string, previousValue: unknown, nextValue: unknown) {
+      if (this._config.values.debug) {
+        if (
+          nextValue !== previousValue &&
+          typeof nextValue === 'object' &&
+          typeof previousValue === 'object' &&
+          JSON.stringify(nextValue) === JSON.stringify(previousValue)
+        ) {
+          this._log.warn(`Option "${key}" value is the same as the previous one but the reference is different`);
+          this._log.warn(`You should avoid changing the reference of the object to prevent unnecessary calculations`);
+          this._log.warn(`"${key}" previous value:`, previousValue);
+          this._log.warn(`"${key}" new value:`, nextValue);
+        }
+      }
+    }
+
+    private _processCustomConfigs(pluginManager: PluginController): void {
+      const customConfigs = pluginManager.configRegistry.getAll();
+
+      // Rebuild the custom attribute mapping and names set
+      this._customAttrKeyMapping = {};
+      this._customConfigKeys = new Set(customConfigs.keys());
+
+      // Clean up subscriptions for configs that no longer exist
+      for (const [name, unsub] of this._customConfigSubscriptions) {
+        if (!customConfigs.has(name)) {
+          unsub();
+          this._customConfigSubscriptions.delete(name);
+        }
+      }
+
+      for (const [name, definition] of customConfigs) {
+        // Build attribute name mappings (kebab-case and lowercase)
+        // Add to mapping unless attribute is explicitly disabled (default is true)
+        if (definition.attribute) {
+          const attrNames = this._getAttributeNames(name);
+          for (const attrName of attrNames) {
+            this._customAttrKeyMapping[attrName] = name;
+          }
+        }
+
+        let preExistingValue: unknown;
+        let hasPreExistingValue = false;
+
+        const existingDescriptor = Object.getOwnPropertyDescriptor(this, name);
+        const isExistingDataDescriptor = !!existingDescriptor && !existingDescriptor.get && !existingDescriptor.set;
+        if (isExistingDataDescriptor && existingDescriptor.value !== undefined) {
+          preExistingValue = existingDescriptor.value;
+          hasPreExistingValue = true;
+        }
+
+        if (!hasPreExistingValue && !!definition.attribute) {
+          for (const attrName of this._getAttributeNames(name)) {
+            const attrValue = this.getAttribute(attrName);
+            if (attrValue !== undefined && attrValue !== null) {
+              try {
+                preExistingValue = definition.fromAttribute ? definition.fromAttribute(attrValue) : attrValue;
+              } catch (error) {
+                this._log.warn(`fromAttribute() for "${name}" threw an error, using raw attribute value`, error);
+                preExistingValue = attrValue;
+              }
+              hasPreExistingValue = true;
+              break;
+            }
+          }
+        }
+
+        // No separate state seeding here: `buildPluginApi`'s `registerConfig`
+        // already seeded this key on the ctx's `ConfigController` (via
+        // `ConfigController.register`) at plugin-setup time, before this
+        // definition could ever appear in `configRegistry.getAll()`. The
+        // registry lookup below is used only for adapter metadata (attribute
+        // mapping, `fromAttribute`/`normalize`), not to re-seed state.
+
+        // Create property accessor (getter/setter) if not already defined
+        const descriptor = Object.getOwnPropertyDescriptor(this, name);
+        if (!descriptor || !descriptor.set || !descriptor.get) {
+          Object.defineProperty(this, name, {
+            set: (value: unknown) => {
+              // Use _setValue for consistent handling (normalization happens inside _setValue)
+              this._setValue(name, value);
+            },
+            get: () => {
+              // Use _getValue for consistent handling
+              return this._getValue(name);
+            },
+            enumerable: true,
+            configurable: true,
+          });
+        }
+
+        // Subscribe to state changes (only if not already subscribed)
+        if (!this._customConfigSubscriptions.has(name)) {
+          // Atomic per-key `observeCustom` — same contract as v1's
+          // `this.sub(stateKey, cb, false)` (init=false: no immediate call, only on
+          // subsequent changes). `_setValue`'s early-return guard prevents circular
+          // updates.
+          const unsub = this._config.observeCustom(name, (nextValue) => {
+            this._setValue(name, nextValue);
+          });
+          this._customConfigSubscriptions.set(name, unsub);
+          this.addDisposer(unsub);
+        }
+
+        if (hasPreExistingValue) {
+          this._setValue(name, preExistingValue);
+          // Same re-adoption hazard as the built-in `allConfigKeys` loop above:
+          // force the flush through in case `_setValue`'s local-cache dedup
+          // no-op'd against a value already cached from a previous ctx.
+          this._flushValueToState(name, this._getValue(name));
+        }
+      }
+    }
+
+    /**
+     * Change-log observer (verbose/debug-gated): logs every `ConfigController`
+     * change as `key: <old> → <new>` — decoupled from the setter and covering
+     * ALL writers (this block, `api`, plugin `registerConfig`), not just
+     * `_setValue`. Recovers the per-key old/new by diffing a snapshot over the
+     * coarse `subscribe` notify (same pattern as the per-key sync subscriptions).
+     * Seeded before the initial-value flush so startup sets are logged; re-seeded
+     * per adoption and torn down via `addDisposer`.
+     */
+    private _setupChangeLog(): void {
+      let snapshot: Record<string, unknown> = { ...this._config.values };
+      this.addDisposer(
+        this._config.subscribe(() => {
+          const next = this._config.values as Record<string, unknown>;
+          for (const key of Object.keys(next)) {
+            const prev = snapshot[key];
+            const curr = next[key];
+            if (!Object.is(curr, prev)) {
+              // Thunked so the `formatConfigLogValue` (JSON.stringify) only runs
+              // when debug is on. Values are quoted for consistency and so an
+              // empty string is visible as `""`.
+              this._log.debug(lazy(() => [`${key}: ${formatConfigLogValue(prev)} → ${formatConfigLogValue(curr)}`]));
+            }
+          }
+          snapshot = { ...next };
+        }),
+      );
+    }
+
+    private _setupCustomConfigs(): void {
+      // Wait for the plugin manager via the `PluginManagerBridge` token: fires
+      // synchronously if the bridge is already resolved (uploader scope attached),
+      // else on the `ensurePluginManager` `get` that constructs it. When it isn't
+      // resolved yet, `whenController` returns a real unsubscriber — `addDisposer`'d
+      // so a re-adoption (ctx-name switch) tears down the previous ctx's pending
+      // waiter instead of leaving it to fire against the wrong controller. (A
+      // synchronous resolve returns a no-op unsub; tracking it is harmless.) In an
+      // editor-alone ctx the bridge is never bound, so this stays inert (no plugins
+      // standalone).
+      this.addDisposer(
+        this.container.whenController(PluginManagerBridge, (bridge) => {
+          const pluginManager = bridge.getPluginManager();
+          // Initial setup.
+          this._processCustomConfigs(pluginManager);
+
+          // Reload custom configs when plugins change. RETURN the unsubscriber as
+          // the waiter's teardown: `whenController`'s own unsub (the `addDisposer`'d
+          // one above) disposes it, so `ChildBlock._releaseController` tears the
+          // plugin-change listener down on release/re-adoption automatically — no
+          // manual `_pluginChangeUnsubscribe` bookkeeping and no teardown-first dance.
+          return pluginManager.onPluginsChange(() => {
+            this._processCustomConfigs(pluginManager);
+          });
+        }),
+      );
+    }
+
+    private _setupMutationObserver(): void {
+      // Create a MutationObserver to detect dynamic attribute changes
+      // This is specifically for custom config attributes that can't be
+      // statically defined in observedAttributes. Built-in attributes are
+      // already handled by the native attributeChangedCallback mechanism.
+      this._mutationObserver = new MutationObserver((mutations) => {
+        for (const mutation of mutations) {
+          if (mutation.type === 'attributes' && mutation.attributeName) {
+            const attrName = mutation.attributeName;
+            const oldValue = mutation.oldValue;
+            const newValue = this.getAttribute(attrName);
+
+            // Skip if value hasn't actually changed
+            if (oldValue === newValue) continue;
+
+            // Skip built-in config attributes - they're handled by observedAttributes
+            const isBuiltInAttr = attrName in builtinAttrKeyMapping;
+            if (isBuiltInAttr) continue;
+
+            // Check if it's a custom plugin config attribute using the mapping
+            if (attrName in this._customAttrKeyMapping) {
+              const key = this._customAttrKeyMapping[attrName] as string;
+              const config = this._getCustomConfigDefinition(key);
+
+              // Call attributeChangedCallback for custom plugin attributes
+              this.attributeChangedCallback(attrName, oldValue ?? '', newValue ?? config?.defaultValue ?? '');
+            }
+          }
+        }
+      });
+
+      // Start observing attribute changes
+      this._mutationObserver.observe(this, {
+        attributes: true,
+        attributeOldValue: true,
+      });
+    }
+
+    /**
+     * Fires on every controller adoption — the initial one and any re-adoption
+     * (ctx-name switch, or ctx death + re-adopt on a v1-managed ctx). All
+     * subscriptions below route through `addDisposer` (or the manual
+     * `ConfigController.subscribe` + dedup, tracked the same way) so a
+     * re-adoption tears the previous cycle's subscriptions down instead of
+     * stacking a second set on top. The plugin-change listener and the
+     * MutationObserver are host/DOM-level and not covered by `addDisposer` — see
+     * the teardown-before-resubscribe below and the idempotent guard,
+     * respectively.
+     */
+    protected override controllerReady(_container: ControllerContainer): void {
+      const anyThis = this as any;
+
+      // Register as this ctx's config writer + warn (deferred) if another host is
+      // already writing config here (one config host per ctx is the contract).
+      // Capture the controller so release/microtask paths don't hit the
+      // post-CONTAINER-clear `@inject` throw.
+      const config = this._config;
+      config.registerWriter(this);
+      this._writerConfig = config;
+      this._maybeWarnMultipleWriters(config);
+
+      // Setup custom configs. The plugin-change listener it opens is torn down on
+      // release (via the `addDisposer`'d `whenController` teardown), so a
+      // re-adoption — which releases before re-adopting — starts fresh instead of
+      // stacking a second listener; no teardown-first needed here.
+      this._setupCustomConfigs();
+
+      // Setup MutationObserver to detect dynamic attribute changes. This
+      // observes the DOM node itself, not the controller/ctx, so it must only
+      // be created once per element lifetime — re-adoption must not stack a
+      // second observer on the same node.
+      if (!this._mutationObserver) {
+        this._setupMutationObserver();
+      }
+
+      // Change-log observer — set up BEFORE the initial-value flush below so the
+      // startup config values are logged too. Verbose/debug-gated.
+      this._setupChangeLog();
+
+      for (const key of allConfigKeys) {
+        // Flush the initial value to the state.
+        // Initial value is taken from the DOM property if it was set before the element was initialized.
+        // If no DOM property was set, the initial value is taken from the initialConfig.
+        const initialValue = anyThis[key] ?? this._config.get(key);
+        if (initialValue !== initialConfig[key]) {
+          this._setValue(key, initialValue);
+          // `_setValue`'s no-op guard (skip when the local property cache
+          // already equals the normalized value) exists to avoid redundant
+          // local writes — it must not also skip seeding a *freshly-adopted*
+          // controller's state on re-adoption (ctx-name switch), whose
+          // `ConfigController` starts from `initialConfig` regardless of what
+          // this element's local cache already holds. Force the flush using
+          // the now-current (possibly already-cached) normalized value.
+          this._flushValueToState(key, anyThis[getLocalPropName(key)]);
+        }
+
+        // Define DOM property setters and getters
+        // They will be used in the userland directly or by the frameworks
+
+        const descriptor = Object.getOwnPropertyDescriptor(this, key);
+        if (!descriptor || !descriptor.set || !descriptor.get) {
+          Object.defineProperty(this, key, {
+            set: (value: unknown) => {
+              this._setValue(key, value);
+            },
+            get: () => {
+              return this._getValue(key);
+            },
+          });
+        }
+      }
+    }
+
+    // Sync each plain (attribute-representable) config key into the local property
+    // on change. Per-key `ConfigController.observe` (atomic dedup) replaces the
+    // coarse `subscribe` + manual `Object.is` loop; change-only (no eager) — the
+    // initial values are seeded by the `allConfigKeys` flush in `controllerReady`.
+    @subscription()
+    protected _wirePlainConfigSync(): Unsubscribe[] {
+      return plainConfigKeys.map((key) => this._config.observe(key, (value) => this._setValue(key, value)));
+    }
+
+    // Computed properties (`cdnCname`, `cameraModes`, …) recompute from their
+    // dependency config keys. Per-key `ConfigController.observe` reproduces the
+    // former per-key `subConfigValue` (dedup, no unrelated-key re-fire); the eager
+    // pass computes the initial value. Composed into one teardown, auto-disposed.
+    @subscription()
+    protected _wireComputedProperties(): Unsubscribe[] {
+      const runComputeProperty = (key: keyof ConfigType) => {
+        computeProperty({
+          key,
+          setValue: this._setValue.bind(this),
+          getValue: this._getValue.bind(this),
+          computationControllers: this._computationControllers,
+        });
+      };
+      return allConfigKeys.map((key) => this._config.observe(key, () => runComputeProperty(key), { immediate: true }));
+    }
+
+    /**
+     * Deferred-confirm warning: one config host per ctx is the contract. On a
+     * register-time conflict, re-check on a microtask and warn only if ≥2 hosts
+     * are still registered AND connected — so a normal host swap (remove-then-add,
+     * reorder), where the outgoing host deregisters a tick later, stays silent.
+     */
+    private _maybeWarnMultipleWriters(config: ConfigController): void {
+      if (config.getWriters().length <= 1) return;
+      queueMicrotask(() => {
+        const live = config.getWriters().filter((w) => w.isConnected);
+        if (live.length > 1 && live.includes(this)) {
+          this._log.warn(
+            `multiple config writers in ctx "${this.containerOrNull?.ctxName ?? '?'}" — only one config host should exist per ctx`,
+          );
+        }
+      });
+    }
+
+    /**
+     * Release counterpart of `controllerReady` (disconnect, or a scope switch
+     * that drops the controller ahead of a re-adopt). Deregisters this writer and
+     * clears the custom-config bookkeeping so a subsequent `controllerReady`
+     * (re-adoption onto a different ctx) starts subscribing fresh instead of
+     * skipping names it thinks are already subscribed on a now-defunct
+     * controller's `ConfigController`. The subscriptions themselves — including
+     * the plugin-change listener returned by the `_setupCustomConfigs` waiter —
+     * are `addDisposer`-registered and already torn down by
+     * `ChildBlock._releaseController` before this hook runs.
+     */
+    protected override controllerReleased(): void {
+      this._writerConfig?.unregisterWriter(this);
+      this._writerConfig = undefined;
+      this._customConfigSubscriptions.clear();
+    }
+
+    public override attributeChangedCallback(name: string, oldVal: string, newVal: string) {
+      super.attributeChangedCallback(name, oldVal, newVal);
+
+      if (oldVal === newVal) return;
+
+      const anyThis = this as any;
+      const builtInKey = builtinAttrKeyMapping[name];
+
+      // Handle built-in config attributes
+      if (builtInKey) {
+        // attributeChangedCallback could be called before the controller is adopted
+        // so we set the DOM property instead of calling this._setValue.
+        // If the block was initialized, the value will be handled by the setter.
+        // If the block was not initialized, the value will be set to the DOM property
+        // and handled on initialization.
+        anyThis[builtInKey] = newVal;
+      } else {
+        // Handle custom config attributes (registered by plugins).
+        //
+        // `attributeChangedCallback` can fire before this element has adopted a
+        // container — during custom-element upgrade (no ctx yet), OR on a live
+        // `ctx-name` switch (the previous container was released and the new one
+        // isn't adopted yet). Guard on `containerOrNull` (never throws) and defer
+        // when it's null: the value is on the DOM attribute and `controllerReady`
+        // → `_setupCustomConfigs` → `_processCustomConfigs` reads pre-existing
+        // attribute values on adoption (matching v1's deferral). The bridge waiter
+        // then resolves the plugin manager now-or-when-available.
+        const container = this.containerOrNull;
+        if (!container) {
+          return;
+        }
+        container.whenController(PluginManagerBridge, (bridge) => {
+          const pluginManager = bridge.getPluginManager();
+          const currentAttrValue = this.getAttribute(name);
+          if (currentAttrValue && currentAttrValue !== newVal) {
+            return;
+          }
+          const key = this._customAttrKeyMapping[name];
+          const config = key ? pluginManager.configRegistry.get(key) : undefined;
+          if (key && config) {
+            // Use fromAttribute to deserialize the value if provided
+            const val = config.fromAttribute ? config.fromAttribute(newVal) : newVal;
+            if (this._getValue(key) === val) return;
+            // Use _setValue for consistent handling (normalization happens inside _setValue)
+            this._setValue(key, val);
+          }
+        });
+      }
+    }
+
+    public override disconnectedCallback(): void {
+      super.disconnectedCallback();
+
+      // `super.disconnectedCallback()` (→ `ChildBlock._releaseController`) already
+      // tore down every `addDisposer`'d subscription — the plugin-change listener
+      // AND the per-custom-key `observeCustom` subs (each `addDisposer`'d in
+      // `_processCustomConfigs`) — and ran `controllerReleased`, which clears
+      // `_customConfigSubscriptions`. Only the MutationObserver needs manual
+      // cleanup: it observes this element's own node (per-element lifetime, created
+      // once behind an idempotency guard), so it's outside the adoption-scoped
+      // disposer engine.
+      if (this._mutationObserver) {
+        this._mutationObserver.disconnect();
+        this._mutationObserver = undefined;
+      }
+    }
+
+    public static get observedAttributes(): string[] {
+      const inherited = (Base as unknown as { observedAttributes?: readonly string[] }).observedAttributes ?? [];
+      const builtInAttrs = Object.keys(builtinAttrKeyMapping);
+
+      // Note: Custom config attributes cannot be statically determined here
+      // since they're registered at runtime. They're handled via mutation observer instead.
+      return [...inherited, ...builtInAttrs];
+    }
+  }
+
+  /**
+   * Define empty DOM properties for all config keys on the mixin class prototype
+   * to make them checkable using `key in element` syntax. This is required for
+   * the frameworks' DOM property bindings to work.
+   */
+  for (const key of allConfigKeys) {
+    (WithConfigClass.prototype as unknown as Record<string, unknown>)[key] = undefined;
+  }
+
+  return WithConfigClass as unknown as T & Constructor<ConfigHost>;
+}
