+++ src/types/exported.ts
@@ -8,7 +8,7 @@ import type {
   FuncFileValidator,
 } from '../abstract/validatorTypes';
 import type { CameraMode } from '../blocks/CameraSource/CameraSource';
-import type { complexConfigKeys } from '../blocks/Config/Config';
+import type { complexConfigKeys } from '../blocks/Config/config-keys';
 import type { FilesViewMode } from '../blocks/UploadList/UploadList';
 
 export {
